Implement live reloading
This change adds a flag to the main program -dev which triggers a reload
when a change to a file is detected in the project directory. It will
recompile and rebuild the generated Pushup app, and send a message to
any connected browsers to reload the page they are on to see the change.