Attempt to clean up context usage
There is somewhat awkward and complicated context package usage in
Pushup. This is because of the needs of live reloading. We want to
distinguish between filesystem change events and OS signal
notifications. Both should propagate normal cancellation expected from
the context package. But filesystem changes should trigger a reload of
the build-and-run process. And trapping OS signals should exit the
program gracefully. Bare context doesn't give us the ability to
detect what the source of cancellation was. So in Pushup we wrap
multiple contexts and implement the context.Context interface, extending
it with some methods and types to get us what we need.

Perhaps a refactoring would approach things differently. Maybe instead
of relying on context.Context for the heavy lifting, distinguishing
between program reload/termination events would start with a different
root object. Only when it is close to the low-level parts of running the
program would context.Context come in to play.