Add dynamic path components
If you add a file where the filename starts with a '$' followed by some
characters, then the resulting Pushup page is available at a route with a
dynamic path. For example, the file <project dir>/pages/x/$name.pushup
is routed via a regular expression like "/x/([^/]+)", which captures the
characters after the /x/ up to the next / (none in this case). "name"
therefore acts like a URL slug from other frameworks. In the Pushup
page, the actual parameter requested by the user is available with
getParam(req, "name") in this example (replace "name" with the URL slug
in the general case).