Add <text></text> to Pushup markup
Inside a keyword block like @if or @for, only a single HTML element may
be expressed, so if multiple elements are desired to be rendered, they
will need to be wrapped in a <text> element. The <text> tags are not
actually emitted, only the element's children, so the HTML semantics
should not be affected.

Also add the htmx active search example.