Add sections
This change add sections to Pushup. Sections are places in layouts that
pages can define what content should appear there. Think common parts of
a web page that individual pages would want to set their own content
for, like page titles, sidebars, and menu navigation.

The implementation of sections adds syntax to Pushup for pages, and a
set of API methods for layouts. See README.md for details.

This change required changing how layouts are rendered, including the
signature of the Respond() method. In a nutshell, all page-related
content, including the main "contents", are rendered concurrently and
the layout synchronizing on each as it encounters a section during
rendering.

A sketch at a Pushup API is also introduced. It relies on the naming of
the receiver of page and layout structs. This is probably not a
long-term solution because it does not allow for information hiding.

Lastly, a `-codegen-only` flag is added to the build and run commands.