New generated type names
This change reworks how type names for the Go struct that houses the
generated code and associated methods for each page are produced.

Previously, they looked like this: for a file named "about.pushup", the
type name would be "Pushup__about__2", where the number "2" would be a
sequential integer to prevent collisions.

Now this page would get the type name "AboutPage", which looks better.

To deal with collisions, the full path is taken in to account and
becomes part of the name, so "foo/bar.pushup" and "quux/bar.pushup"
would produce "FooBarPage" and "QuuxBarPage", respectively.

Layouts get a "Layout" suffix instead of "Page".