Add support for inline partials
Inline partials are blocks denoted with `^partial` within Pushup pages
that can be returned as a response to a client request independent of
the rest of the enclosing page, including any layout.

In addition to the new syntax and runtime support, handling for dev
reload mode was added, such that partials do not get injected with the
reloaded JS.

The end-to-end Pushup server tests in `testdata` were augmented to allow
for multiple differently configured requests per page under the test
harness. This allows, for example, a partial to be tested, in the
context of its page as one request, by itself in another, and any nested
partials in more requests.

The example project in `example` was updated with a few demos of the
inline partial support.

Closes #33.