Refactor codegen (#40)
Code generation is some of the earliest Pushup code, and it was starting
to strain under new feature development and changes. This refactor splits
layouts and pages from each other in terms of code generation, and fixes
some bugs related to that.

A consequence of this change is that partials are now their own type and
implement the Responder interface. This greatly simplifies the logic for
implementing inline partials and not overly complicating the main page
code gen.

This change is not an end of itself, it is intended that it sets the stage for
additional bug fixes, such as with inline partials and control flow statements.

One somewhat major-ish change this commit makes is adding the notion of a
"project file" (the `projectFile` type), which combines both the path to a
Pushup .up file and its subdir in the project directory layout, if any. This
simplifies a lot of logic especially when handling project builds vs tests.

Other changes:

  * Better panic handling in goroutines in server responders
  * Fix sections vis a vis concurrency
  * Reorganize related code to be nearer each other in source