language: Rename ^else and add ^else ^if
This changes the spelling of the keyword denoting an alternate block in
an ^if statement from `else` to `^else`. Note the presence of the
leading transition character. This makes the language more consistent.

This commit also adds new functionality in the form of `^else ^if`. This
allows for more test clauses and expands the expressiveness of the
language.