tests: fix bug in tests with socket file path len
On macOS (according to `man 4 unix`), "UNIX-domain addresses are
variable-length filesystem pathnames of at most 104 characters". The
test code would generate a path for the socket file that would exceed
this maximum length. This is because os.MkdirTemp() uses the $TMPDIR env
var, which on macOS can be quite lenghty (it's not just /tmp, but a
per-user, randomly-named folder under /var/folders/...).