parser: Fix bugs with ^for statements
This change corrects ^for behavior in the parser and analyzer, particularly
that ^for couldn't (erroneously) appear at the top-level of a Pushup doc. It
also makes the Pushup parser more permissive with the structure of statement
blocks; namely, that they can contain multiple HTML elements (not merely a
single one with a nested structure), and Go code blocks (^{...}).